requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
einops>=0.6.0
editdistance>=0.6.3
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.7.0
python-docx>=0.8.11
tqdm>=4.65.0
pyyaml>=6.0
